We want to use the same version liveshare use.

+++ src/services/containers/VsClk.SignalService/RelayServiceHub/RelayHub.cs
@@ -184,7 +184,7 @@ namespace Microsoft.VsCloudKernel.SignalService
                 Id,
                 participantId,
                 properties,
-                changeType,
+                (int)changeType,
                 cancellationToken)));
         }
